minus changes - delete list will trigger delete all tasks function.

+++ ToDoLists/src/main/java/edu/uwb/css533/service/db/ListServiceDB.java
@@ -353,6 +353,12 @@ public class ListServiceDB extends DatabaseConnection{
             if (!accessRemoved) {
                 return removeAccessMsg;
             }
+            // delete all tasks under such list
+            TaskServiceDB taskServiceDB = new TaskServiceDB();
+            String taskDeletion = taskServiceDB.deleteAllTasks(username, listid);
+            if (!taskDeletion.contains("Successfully")) {
+                return taskDeletion;
+            }
 
             String sql = "DELETE FROM Lists WHERE listid=?;";
